[feat/#50] 4계층 분리
