+++ backend/domain/entities/inventory/Medicine.ts
@@ -1,0 +1,6 @@
+export interface Medicine {
+  item_seq: string;
+  item_name: string;
+  entp_name: string | null;
+  type_name: string | null;
+}
